fix(Fluere): replace @Profiles with @ConditionalOnProperty in whole project, add properties in app properties, refactor JwtSecurity.
