+++ docker/docker-compose.yml
@@ -6,7 +6,6 @@ services:
       - "27017:27017"
     volumes:
       - mongo_dev:/data/db
-      - ./init-mongo.js:/docker-entrypoint-initdb.d/init-mongo.js
     environment:
       MONGO_INITDB_ROOT_USERNAME: admin
       MONGO_INITDB_ROOT_PASSWORD: admin
