+++ src/main/java/com/example/fluere/bootloader/MongoService.java
@@ -1,20 +1,18 @@
 package com.example.fluere.bootloader;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Profile;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.stereotype.Service;
 
 @Service
-@Profile({"dev","test"})
 @RequiredArgsConstructor
+@ConditionalOnProperty(name = "app.property.service.bootloader-info", havingValue = "true")
 public class MongoService {
     private final MongoTemplate mongoTemplate;
 
-    //TODO RIGHT NOW LET IT BE BUT LATER REMOVE THIS CLASS FOR DB CONNECTION TEST
     public void performConnectivityCheck() {
         try {
-            // Try a simple operation, e.g., listing collections
             mongoTemplate.getCollectionNames().forEach(System.out::println);
             System.out.println("MongoDB connectivity check successful.");
         } catch (Exception e) {
