+++ src/main/java/com/example/fluere/bootloader/config/ServletInitializerDev.java
@@ -1,16 +1,15 @@
 package com.example.fluere.bootloader.config;
 
 import com.example.fluere.FluereApplicationDev;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-import org.springframework.context.annotation.Profile;
 
-@Profile({"dev", "test"})
-//TODO change this config for the profiles. Separate for test, dev and prod
+@ConditionalOnProperty(name = "app.property.service.bootloader-info", havingValue = "true")
 public class ServletInitializerDev extends SpringBootServletInitializer {
 
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(FluereApplicationDev.class);
-	}
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+        return application.sources(FluereApplicationDev.class);
+    }
 }
