+++ src/main/java/com/example/fluere/userentity/service/admin/DefaultAdminUserService.java
@@ -6,8 +6,8 @@ import com.example.fluere.budget.service.admin.AdminBudgetService;
 import com.example.fluere.exception.custom.UnableToCreateException;
 import com.example.fluere.security.request.AuthenticationRequest;
 import com.example.fluere.security.request.RegisterUserRequest;
-import com.example.fluere.security.service.AuthenticationService;
-import com.example.fluere.security.service.JwtService;
+import com.example.fluere.security.service.jwt.AuthenticationService;
+import com.example.fluere.security.service.jwt.JwtService;
 import com.example.fluere.userentity.controller.admin.AdminController;
 import com.example.fluere.userentity.model.AppUser;
 import com.example.fluere.userentity.model.Role;
