+++ src/main/java/com/example/fluere/userentity/service/user/DefaultUserService.java
@@ -7,8 +7,8 @@ import com.example.fluere.exception.custom.UnableToCreateException;
 import com.example.fluere.security.request.AuthenticationRequest;
 import com.example.fluere.security.request.RegisterUserRequest;
 import com.example.fluere.security.response.RegisterResponseDTO;
-import com.example.fluere.security.service.AuthenticationService;
-import com.example.fluere.security.service.JwtService;
+import com.example.fluere.security.service.jwt.AuthenticationService;
+import com.example.fluere.security.service.jwt.JwtService;
 import com.example.fluere.userentity.controller.user.AppUserController;
 import com.example.fluere.userentity.model.AppUser;
 import com.example.fluere.userentity.model.UserIdWrapper;
